requirements.txt
Flask
flask-cors
google-generativeai
pymongo
vaderSentiment
gunicorn
flask_sqlalchemy